+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added custom color support
 * Added space support
 * Added custom space support
+* Added support for space between widgets in a list of widgets like Row and Column
 
 ## 0.0.5 - 2024-10-16